refactor: remove legacy format conversion now that it's unused
