+++ lib/index.js
@@ -1,3 +1,3 @@
-const { getDoc: getOpenapiDoc } = require('./openapi')
+const getOpenapiDoc = require('./get-openapi-doc')
 
 module.exports = getOpenapiDoc
